Restart Makefile
Actually, this project will have more than one Makefile, reflecting different
stages of development and usage of django-project-skeleton.

1) The Makefile, that is included in the main directory when you download/use
the repo as Django's startproject template:
This Makefile is used during development and primarily includes receipts to
support development.
However, this will be the Makefile just after downloading the skeleton. So it
has to include a receipt to prepare the project for further usage.

2) The Makefile, that is used during the lifespan of the project:
This Makefile should support project-specific development, testing and
deployment until production usage.