Makefile supports Docker build step
The actual deployment Makefile received a lot of love, including
- comments, comments, documentation
- switched from 'echo' to '/usr/bin/printf' for max compatibility
- some minor manual testing of Makefile commands/targets

docker-compose.yml was updated to include the new naming scheme for images
built through the Makefile interface.
The tagging may be controlled by Makefile variables, that also are set up to
support templating by Django's startproject.
To tag every build with an unique identifier, a custom shell script was added
in the new ./bin directory. This script needs documentation and comments.