Building Docker images using the tox environment
Success! The build process can now be launched directly from within this repo,
no more need to use a seperate test bed.

The current structure is deployed into a temporary tox environment,
initialized and then the Docker build process is triggered.

Images will be tagged with the sha1 tags of this repo.