Got stuck with Django's logging inside of the container
It seems like everything is working fine, but I can't see the logs in the
Docker container, if I try to manually inject them from a second TTY using
Django's shell and pure Python.

TODO: Write a test function bound to a url to test if logging can be triggered
from the outside.