Synchronize time between host and containers
This is not working as expected for the Django container. The first messages
are logged with correct timestamps (these are issued by Gunicorn), but after
Django successfully applied the provided configuration, the time zone for all
Python-based logs is set to the value of Django's TIME_ZONE (which, if not
specified, is set to 'America/Chicago').

TODO: Fix this issue; do **not** rely on Django's TIME_ZONE for the
container's logging timestamps. Probably an own implementation may be
necessary, to unify logging timestamps throughout all containers.
However, this is highly necessary for any debugging, especially if any log
aggregator is used.